% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate}
\alias{annotate}
\title{Annotate}
\usage{
annotate(data, anno_env, col_spec)
}
\arguments{
\item{data}{Performance data}

\item{anno_env}{Environment containing the annotation functions.}

\item{col_spec}{Lists of id_cols, ordering_cols, perf_cols}
}
\description{
Runn annotation functions over
}
\seealso{
source_annotations
}
