% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_configuration.r
\name{build_configuration}
\alias{build_configuration}
\title{Build Configuration}
\usage{
build_configuration(path_to_config)
}
\description{
Build runtime configuration by reading configuration from path or use defaults.
}
