% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{digestion}
\alias{digestion}
\title{Digestion}
\usage{
digestion(config)
}
\arguments{
\item{config}{Runtime configuration}
}
\value{
table of performers and their annotations
}
\details{
The config parameter contains
}
