% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_annotations.r
\name{distill_annotations}
\alias{distill_annotations}
\title{Distill Annotations}
\usage{
distill_annotations(annotations, uri_lookup)
}
\description{
Prepare annotations for jsonld export: 
  Filter for annotations that are TRUE,
  Rename columns for export to jsonld,
  Convert annoation short names to full url
}
