% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_in_uris.r
\name{lookup_new_name}
\alias{lookup_new_name}
\title{lookup new name}
\usage{
lookup_new_name(x, lu_list)
}
\arguments{
\item{x}{name to lookup}

\item{lu_list}{lookup list to check x against names for value of new name}
}
\value{
new name or original x if x not present in names of lookup list.
}
