% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{main}
\alias{main}
\title{Main}
\usage{
main(path_to_config = NULL, overrides = NULL)
}
\arguments{
\item{path_to_config}{Path to configuration yaml. Use NULL to use internal defaults.}

\item{overrides}{List of configuration overrides.}
}
\description{
The entry function with all the side effects
}
