% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_performers.r
\name{merge_performers}
\alias{merge_performers}
\title{Merge Performers}
\usage{
merge_performers(spek, performers_table)
}
\description{
Apply annotations to performers in spek.
}
\details{
Adds newly found performers and updates esiting performers in spek
}
