% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_performers.r
\name{mundify_list}
\alias{mundify_list}
\title{Mundify List}
\usage{
mundify_list(x, val, keep.null = FALSE)
}
\description{
Essentially modifyList that also adds unnamed elements if the identical value isn't already present.
}
