% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performers.r
\name{performers}
\alias{performers}
\title{Performers}
\usage{
performers(dispositions, app_onto_url, uri_lookup)
}
\description{
Condense data frame with single row per performer and list of their annotations.
}
\note{
Convenient format for subsequent export to jsonld
}
