% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist_to_disk.r
\name{persist_to_disk}
\alias{persist_to_disk}
\title{Persist to Disk}
\usage{
persist_to_disk(content, output_dir = tempdir())
}
\arguments{
\item{content}{json content}

\item{output_dir}{directory to create and write output into}
}
\description{
Persist json to disk
}
