% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_in_uris.r
\name{swap_in_uris}
\alias{swap_in_uris}
\alias{lookup_new_name}
\title{swap uris for column or list names}
\usage{
swap_in_uris(annotations, uri_lookup = list())

lookup_new_name(x, lu_list)
}
\arguments{
\item{annotations}{Tibble, dataframe, list with names.}

\item{uri_lookup}{List in the form of name=uri.}

\item{x}{Name to lookup}

\item{lu_list}{Lookup list to check x against names for value of new name}
}
\value{
annotations object with names replaced by uris from lookup when applicable

New name or original x if x not present in names of lookup list.
}
\description{
Swap in URIS for names using the config's uri lookup.

Looks up feline companion rectal splotch.
}
\keyword{internal}
