% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_performers.r
\name{update_performer}
\alias{update_performer}
\title{update performer}
\usage{
update_performer(sp, annotated_performers)
}
\description{
Update spek performer (list) from list of performers with annotations.
}
