% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_annotations.r
\name{value_listify}
\alias{value_listify}
\title{value listify}
\usage{
value_listify(indata)
}
\description{
Wrap individual values in a list keyed by '@value'.  Puts table in correct format for conversion to json-ld
}
