% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{annotate}
\alias{annotate}
\title{Annotate}
\usage{
annotate(data, anno_env, perf_cols)
}
\arguments{
\item{data}{Performance data}

\item{anno_env}{Environment containing the annotation functions.}

\item{perf_cols}{names of columns of performance data}
}
\description{
Runn annotation functions over
}
\seealso{
source_annotations
}
