% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{main}
\alias{main}
\title{Main}
\usage{
main(config_path = NULL, ...)
}
\arguments{
\item{config_path}{Path to configuration yaml. Use NULL to use internal defaults.}

\item{...}{List of configuration overrides passed to build_config}
}
\description{
The entry function with all the side effects
}
\seealso{
build_configuration
}
