% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_annotations.r
\name{source_annotations}
\alias{source_annotations}
\title{Source Annotations}
\usage{
source_annotations(path)
}
\arguments{
\item{path}{Path too annotations R file.}
}
\description{
Source the annotations.r file which contain the fucntions to apply to the 
  data in order to make inferences about performers from the data.
}
