% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_in_uris.r
\name{swap_in_uris}
\alias{swap_in_uris}
\title{swap uris for column or list names}
\usage{
swap_in_uris(annotations, uri_lookup = list())
}
\arguments{
\item{annotations}{tibble, dataframe, list with names.}

\item{uri_lookup}{list in the form of name=uri.}
}
\value{
annotations object with names replaced by uris from lookup when applicable
}
\description{
Swap in URIS for names using the config's uri lookup
}
